+++ lib/Arduino-Teensy-Codec/play_sd_wav.h
@@ -57,6 +57,10 @@ public:
 	//int play(const size_t p, const size_t size) {stop();if (!fopen(p,size)) return ERR_CODEC_FILE_NOT_FOUND; return play();}
 	//int play(const uint8_t*p, const size_t size) {stop();if (!fopen(p,size))  return ERR_CODEC_FILE_NOT_FOUND; return play();}
 	unsigned lengthMillis(void);
+	unsigned positionMillis(void);
+	unsigned int sampleRate(void) { return samprate; }
+	unsigned int bitResolution(void) { return bitsPerSample; }
+	unsigned int parseHeader(MutexFsBaseFile *file);
 	size_t fposition(void) { return AudioCodec::fposition() - sd_left; }
 
 protected:
@@ -64,14 +68,14 @@ protected:
 	uint8_t			*sd_p;
 	int				sd_left;
 
-	short			*buf[2];
+	int32_t			*buf[2];
 	size_t			decoded_length[2];
 	size_t			decoding_block;
 	unsigned int	decoding_state; //state 0: read sd, state 1: decode
 
 	uintptr_t 		play_pos;
 
-	unsigned short	samprate;
+	unsigned int	samprate;
 	unsigned short	bitsPerSample;
 	size_t			data_size;