+++ lib/Audio/output_i2s.cpp
@@ -61,27 +61,27 @@ static const uint32_t vol_table[101] = {
     65536
 };
 
-void memcpy_tointerleaveLR(int8_t vol, int32_t *dst, const int16_t *srcL, const int16_t *srcR)
+void memcpy_tointerleaveLR(int8_t vol, int32_t *dst, const int32_t *srcL, const int32_t *srcR)
 {
     for (int i = 0; i < AUDIO_BLOCK_SAMPLES/2; i++) {
-        dst[i*2+0] = (int32_t) srcL[i] * vol_table[vol] + DAC_ZERO_OFFSET;
-        dst[i*2+1] = (int32_t) srcR[i] * vol_table[vol] + DAC_ZERO_OFFSET;
+        dst[i*2+0] = (int32_t)(((int64_t)srcL[i] * vol_table[vol]) >> 8) + DAC_ZERO_OFFSET;
+        dst[i*2+1] = (int32_t)(((int64_t)srcR[i] * vol_table[vol]) >> 8) + DAC_ZERO_OFFSET;
     }
 }
 
-void memcpy_tointerleaveL(int8_t vol, int32_t *dst, const int16_t *srcL)
+void memcpy_tointerleaveL(int8_t vol, int32_t *dst, const int32_t *srcL)
 {
     for (int i = 0; i < AUDIO_BLOCK_SAMPLES/2; i++) {
-        dst[i*2+0] = (int32_t) srcL[i] * vol_table[vol] + DAC_ZERO_OFFSET;
+        dst[i*2+0] = (int32_t)(((int64_t)srcL[i] * vol_table[vol]) >> 8) + DAC_ZERO_OFFSET;
         dst[i*2+1] = 0 + DAC_ZERO_OFFSET;
     }
 }
 
-void memcpy_tointerleaveR(int8_t vol, int32_t *dst, const int16_t *srcR)
+void memcpy_tointerleaveR(int8_t vol, int32_t *dst, const int32_t *srcR)
 {
     for (int i = 0; i < AUDIO_BLOCK_SAMPLES/2; i++) {
         dst[i*2+0] = 0 + DAC_ZERO_OFFSET;
-        dst[i*2+1] = (int32_t) srcR[i] * vol_table[vol] + DAC_ZERO_OFFSET;
+        dst[i*2+1] = (int32_t)(((int64_t)srcR[i] * vol_table[vol]) >> 8) + DAC_ZERO_OFFSET;
     }
 }
 
@@ -500,6 +500,26 @@ void AudioOutputI2S::config_i2s(void)
 #endif
 }
 
+
+#if defined(__IMXRT1062__)
+void AudioOutputI2S::setFrequency(float fs)
+{
+	// Reconfigure PLL4 (shared by SAI1/I2S and SAI2/SPDIF)
+	int n1 = 4; //SAI prescaler 4 => (n1*n2) = multiple of 4
+	int n2 = 1 + (24000000 * 27) / ((int)fs * 256 * n1);
+
+	double C = ((double)fs * 256 * n1 * n2) / 24000000;
+	int c0 = C;
+	int c2 = 10000;
+	int c1 = C * c2 - (c0 * c2);
+	set_audioClock(c0, c1, c2, true); // force=true for runtime reconfiguration
+
+	// Update SAI1 clock dividers
+	CCM_CS1CDR = (CCM_CS1CDR & ~(CCM_CS1CDR_SAI1_CLK_PRED_MASK | CCM_CS1CDR_SAI1_CLK_PODF_MASK))
+		   | CCM_CS1CDR_SAI1_CLK_PRED(n1-1) // &0x07
+		   | CCM_CS1CDR_SAI1_CLK_PODF(n2-1); // &0x3f
+}
+#endif
 
 /******************************************************************/